Improved Frontend File Structure
