+++ frontend/src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import LandingPage from './components/LandingPage/LandingPage';
+import LandingPage from './pages/LandingPage';
 
 function App() {
   return (
