+++ frontend/src/assets/images/logo.png
