+++ frontend/src/components/LandingPage/Header.js
@@ -1,5 +1,5 @@
 import React, {useState, useEffect} from 'react'
-import Logo from './../../images/logo.png'
+import Logo from './../../assets/images/logo.png'
 
 function Header() {
